Move the new spec helper methods in testing_support modules
