+++ backend/spec/spec_helper.rb
@@ -36,6 +36,8 @@ require 'spree/testing_support/flash'
 require 'spree/testing_support/url_helpers'
 require 'spree/testing_support/order_walkthrough'
 require 'spree/testing_support/capybara_ext'
+require 'spree/testing_support/precompiled_assets'
+require 'spree/testing_support/translations'
 
 require 'capybara-screenshot/rspec'
 Capybara.save_path = ENV['CIRCLE_ARTIFACTS'] if ENV['CIRCLE_ARTIFACTS']
@@ -59,14 +61,6 @@ Capybara.default_max_wait_time = ENV['DEFAULT_MAX_WAIT_TIME'].to_f if ENV['DEFAU
 
 ActiveJob::Base.queue_adapter = :test
 
-def check_missing_translations(page, example)
-  missing_translations = page.body.scan(/translation missing: #{I18n.locale}\.(.*?)[\s<\"&]/)
-  if missing_translations.any?
-    puts "Found missing translations: #{missing_translations.inspect}"
-    puts "In spec: #{example.location}"
-  end
-end
-
 RSpec.configure do |config|
   config.color = true
   config.infer_spec_type_from_file_location!
@@ -87,21 +81,6 @@ RSpec.configure do |config|
     DatabaseCleaner.clean_with :truncation
   end
 
-  config.when_first_matching_example_defined(type: :feature) do
-    config.before :suite do
-      # Preload assets
-      # This should avoid capybara timeouts, and avoid counting asset compilation
-      # towards the timing of the first feature spec.
-      Rails.application.precompiled_assets
-    end
-  end
-
-  config.when_first_matching_example_defined(type: :system) do
-    config.before :suite do
-      Rails.application.precompiled_assets
-    end
-  end
-
   config.before do
     Rails.cache.clear
     if RSpec.current_example.metadata[:js] && page.driver.browser.respond_to?(:url_blacklist)
@@ -112,14 +91,6 @@ RSpec.configure do |config|
   config.include BaseFeatureHelper, type: :feature
   config.include BaseFeatureHelper, type: :system
 
-  config.after(:each, type: :feature) do |example|
-    check_missing_translations(page, example)
-  end
-
-  config.after(:each, type: :system) do |example|
-    check_missing_translations(page, example)
-  end
-
   config.include FactoryBot::Syntax::Methods
   config.include ActiveSupport::Testing::Assertions
   config.include ActiveJob::TestHelper
@@ -128,6 +99,7 @@ RSpec.configure do |config|
   config.include Spree::TestingSupport::UrlHelpers
   config.include Spree::TestingSupport::ControllerRequests, type: :controller
   config.include Spree::TestingSupport::Flash
+  config.include Spree::TestingSupport::Translations
 
   config.extend WithModel
