+++ frontend/spec/spec_helper.rb
@@ -36,6 +36,8 @@ require 'spree/testing_support/flash'
 require 'spree/testing_support/url_helpers'
 require 'spree/testing_support/order_walkthrough'
 require 'spree/testing_support/caching'
+require 'spree/testing_support/blacklist_urls'
+require 'spree/testing_support/translations'
 
 require 'capybara-screenshot/rspec'
 Capybara.save_path = ENV['CIRCLE_ARTIFACTS'] if ENV['CIRCLE_ARTIFACTS']
@@ -45,20 +47,6 @@ require "selenium/webdriver"
 Capybara.javascript_driver = (ENV['CAPYBARA_DRIVER'] || :selenium_chrome_headless).to_sym
 
 ActiveJob::Base.queue_adapter = :test
-
-def check_missing_translations(page, example)
-  missing_translations = page.body.scan(/translation missing: #{I18n.locale}\.(.*?)[\s<\"&]/)
-  if missing_translations.any?
-    puts "Found missing translations: #{missing_translations.inspect}"
-    puts "In spec: #{example.location}"
-  end
-end
-
-def set_url_blacklist(browser)
-  if browser.respond_to?(:url_blacklist)
-    browser.url_blacklist = ['http://fonts.googleapis.com']
-  end
-end
 
 RSpec.configure do |config|
   config.color = true
@@ -91,28 +79,14 @@ RSpec.configure do |config|
     Rails.cache.clear
   end
 
-  config.before(:each, type: :feature) do
-    set_url_blacklist(page.driver.browser)
-  end
-
-  config.before(:each, type: :system) do
-    set_url_blacklist(page.driver.browser)
-  end
-
-  config.after(:each, type: :feature) do |example|
-    check_missing_translations(page, example)
-  end
-
-  config.after(:each, type: :system) do |example|
-    check_missing_translations(page, example)
-  end
-
   config.include FactoryBot::Syntax::Methods
 
   config.include Spree::TestingSupport::Preferences
   config.include Spree::TestingSupport::UrlHelpers
   config.include Spree::TestingSupport::ControllerRequests, type: :controller
   config.include Spree::TestingSupport::Flash
+  config.include Spree::TestingSupport::BlacklistUrls
+  config.include Spree::TestingSupport::Translations
 
   config.example_status_persistence_file_path = "./spec/examples.txt"
